+++ R/RcppExports.R
@@ -937,14 +937,6 @@ moisture_tissueRWC <- function(psiSym, pi0, epsilon, psiApo, c, d, af, femb = 0.
     .Call(`_medfate_tissueRelativeWaterContent`, psiSym, pi0, epsilon, psiApo, c, d, af, femb)
 }
 
-moisture_cohortFMC <- function(spwb, SpParams) {
-    .Call(`_medfate_cohortFMC`, spwb, SpParams)
-}
-
-moisture_cohortFMCDay <- function(spwb_day, x, SpParams) {
-    .Call(`_medfate_cohortFMCDay`, spwb_day, x, SpParams)
-}
-
 transp_profitMaximization <- function(supplyFunction, photosynthesisFunction, Gswmin, Gswmax, gainModifier = 1.0, costModifier = 1.0, costWater = "dEdP") {
     .Call(`_medfate_profitMaximization`, supplyFunction, photosynthesisFunction, Gswmin, Gswmax, gainModifier, costModifier, costWater)
 }