+++ src/paramutils.cpp
@@ -476,6 +476,15 @@ NumericVector conduit2sapwoodWithImputation(IntegerVector SP, DataFrame SpParams
   }
   return(conduit2sapwood);
 }
+NumericVector maxFMCWithImputation(IntegerVector SP, DataFrame SpParams) {
+  NumericVector maxFMC = speciesNumericParameter(SP, SpParams, "maxFMC");
+  for(int c=0;c<maxFMC.size();c++) {
+    if(NumericVector::is_na(maxFMC[c])) {
+      maxFMC[c] = 120.0; //To be improved
+    }
+  }
+  return(maxFMC);
+}
 NumericVector stemPI0WithImputation(IntegerVector SP, DataFrame SpParams) {
   NumericVector WoodDensity = woodDensityWithImputation(SP, SpParams);
   NumericVector StemPI0 = speciesNumericParameter(SP, SpParams, "StemPI0");
@@ -1272,6 +1281,7 @@ NumericVector speciesNumericParameterWithImputation(IntegerVector SP, DataFrame 
     else if(parName == "SRL") return(specificRootLengthWithImputation(SP, SpParams));
     else if(parName == "RLD") return(rootLengthDensityWithImputation(SP, SpParams));
     else if(parName == "conduit2sapwood") return(conduit2sapwoodWithImputation(SP, SpParams));
+    else if(parName == "maxFMC") return(maxFMCWithImputation(SP, SpParams));
     else if(parName == "StemPI0") return(stemPI0WithImputation(SP, SpParams));
     else if(parName == "StemEPS") return(stemEPSWithImputation(SP, SpParams));
     else if(parName == "StemAF") return(stemAFWithImputation(SP, SpParams));